base, simple gemspec
