+++ objc_dependency_tree_generator.gemspec
@@ -1,0 +1,14 @@
+Gem::Specification.new do |s|
+  s.name        = 'objc_dependency_tree_generator'
+  s.version     = '0.0.1'
+  s.date        = '2015-08-19'
+  s.summary     = 'Objective-C and Swift dependency tree generator'
+  s.description = 'Tool that allows to generate Objective-C and Swift dependency tree from object files'
+  s.authors     = ['Paul Taykalo']
+  s.email       = 'tt.kilew@gmail.com'
+  s.files       = ['lib/objc_dependency_tree_generator.rb']
+  s.homepage    =
+      'http://rubygems.org/gems/hola'
+  s.license       = 'MIT'
+  s.executables << 'objc_dependency_tree_generator'
+end
